Added f32 type and changes from Standards.

+++ lib/stds/src/camera.c
@@ -1,12 +1,44 @@
+//=============================================================================================//
+// FILENAME :       camera.c
+//
+// DESCRIPTION :
+//        This file defines the function for updating a camera around a parent entity. Generally,
+//        this should be the player. All offsets are pre-applied to the draw functions. To keep 
+//        something from being updated, just re-add the valuesx + app.camera.x, y + app.camera.y.
+//
+// PUBLIC FUNCTIONS :
+//        void update_camera( entity_t * );
+//
+// NOTES :
+//        Permission is hereby granted, free of charge, to any person obtaining a copy
+//        of this software and associated documentation files (the "Software"), to deal
+//        in the Software without restriction, including without limitation the rights
+//        to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//        copies of the Software, and to permit persons to whom the Software is
+//        furnished to do so, subject to the following conditions:
+//
+//        The above copyright notice and this permission notice shall be included in all
+//        copies or substantial portions of the Software.
+//
+//        THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//        IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//        FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//        AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//        LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//        OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+//        SOFTWARE.
+//
+// AUTHOR :   Joshua Crotts        START DATE :    18 Jun 2020
+//
+//=============================================================================================//
+
 #include "../include/camera.h"
 
 /**
  * Applies an offset from the supplied entity, and stores
  * the coordinates in the App struct. Typically, in an overhead
  * style game, this supplied entity will be a reference to the
- * player. All offsets are pre-applied to the draw functions.
- * To keep something from being updated, just re-add the values
- * x + app.camera.x, y + app.camera.y.
+ * player. 
  *
  * @param entity_t* pointer to parent entity.
  *
@@ -15,8 +47,8 @@
 void
 update_camera( entity_t *focus_point ) {
   if ( focus_point != NULL ) {
-    app.camera.x = ( ( int32_t ) focus_point->x + focus_point->w / 2 ) - ( app.SCREEN_WIDTH >> 1 );
-    app.camera.y = ( ( int32_t ) focus_point->y + focus_point->h / 2 ) - ( app.SCREEN_HEIGHT >> 1 );
+    app.camera.x = ( focus_point->x + focus_point->w / 2 ) - ( app.SCREEN_WIDTH >> 1 );
+    app.camera.y = ( focus_point->y + focus_point->h / 2 ) - ( app.SCREEN_HEIGHT >> 1 );
     app.camera.w = app.SCREEN_WIDTH;
     app.camera.h = app.SCREEN_HEIGHT;
